requirements.txt
matplotlib==3.10.3
numpy==2.3.1
numpy_financial==1.0.0
pandas==2.3.1
plotly==6.2.0
python_dateutil==2.9.0.post0
streamlit==1.46.1
